WorkProject/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-jsoneditor==0.2.2
packaging==21.3
pkg_resources==0.0.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
